Adding fix for tabular tests

+++ tabular/tests/unittests/test_tabular.py
@@ -137,7 +137,8 @@ def test_advanced_functionality():
     savedir_predictor_original = savedir + "predictor/"
     predictor: TabularPredictor = TabularPredictor(label=label, path=savedir_predictor_original).fit(train_data)
     leaderboard = predictor.leaderboard(data=test_data)
-    predictor.plot_ensemble_model()
+    if os.name != "nt":
+        predictor.plot_ensemble_model()
     extra_metrics = ["accuracy", "roc_auc", "log_loss"]
     test_data_no_label = test_data.drop(columns=[label])
     with pytest.raises(ValueError):
@@ -247,7 +248,8 @@ def test_advanced_functionality():
 
     assert predictor.get_model_full_dict() == dict()
     predictor.refit_full()
-    predictor.plot_ensemble_model()
+    if os.name != "nt":
+        predictor.plot_ensemble_model()
     assert len(predictor.get_model_full_dict()) == num_models
     assert len(predictor.get_model_names()) == num_models * 2
     for model in predictor.get_model_names():
@@ -1143,7 +1145,8 @@ def test_tabular_log_to_file():
     predictor = TabularPredictor(label="class", log_to_file=True, log_file_path=log_file).fit(train_data=train_data, hyperparameters={"DUMMY": {}})
     log = TabularPredictor.load_log(log_file_path=log_file)
     assert "TabularPredictor saved." in log[-1]
-    os.remove(log_file)
+    if os.name != "nt":
+        os.remove(log_file)
 
     with pytest.raises(AssertionError):
         TabularPredictor.load_log()
